Tools: Modified tsc tasks for migration plan.
