--- ts/global.d.ts
@@ -1,10 +1,0 @@
-declare module 'highcharts' {
-    const seriesTypes: any;
-    type ColorString = string;
-    type GradientColorObject = any;
-    function isNumber (...args: Array<any>): any;
-    function relativeLength (...args: Array<any>): any;
-    function merge (...args: Array<any>): any;
-    function pick (...args: Array<any>): any;
-    function seriesType (...args: Array<any>): any;
-}
